+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M24P800FD2DV630G0BV2ER03"),
+  id: z.string().default("01M25CQY93ZMN7C1A9ZF47W352"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>